[#65] refactor : 파일명 변경, 임시 렌더링 제거
