+++ components/Layout/MobileFooterNavigationBar/MobileFooterNavigationBar.tsx
@@ -1,18 +1,18 @@
 import Link from "next/link";
-import * as styles from "./FooterNavigationBar.css";
+import * as styles from "./MobileFooterNavigationBar.css";
 
-const FOOTER_NAVIGATION_LIST = [
+const MOBILE_FOOTER_NAVIGATION = [
   { id: 1, name: "홈", icon: "", href: "/home" },
   { id: 2, name: "육아일기", icon: "", href: "/diary" },
   { id: 3, name: "건강수첩", icon: "", href: "/healthlog" },
   { id: 4, name: "관리", icon: "", href: "settings" },
 ];
 
-const FooterNavigationBar = () => {
+const MobileFooterNavigationBar = () => {
   return (
     <footer className={styles.footer}>
       <nav className={styles.nav}>
-        {FOOTER_NAVIGATION_LIST.map((menu) => (
+        {MOBILE_FOOTER_NAVIGATION.map((menu) => (
           <Link className={styles.link} key={menu.id} href={menu.href}>
             {menu.name}
           </Link>
@@ -22,4 +22,4 @@ const FooterNavigationBar = () => {
   );
 };
 
-export default FooterNavigationBar;
+export default MobileFooterNavigationBar;
