+++ components/Layout/MobileHeaderLogo/MobileHeaderLogo.tsx
@@ -1,8 +1,8 @@
 import Image from "next/image";
-import * as styles from "./HeaderLogo.css";
+import * as styles from "./MobileHeaderLogo.css";
 import sampleImageSrc from "@/assets/edit.svg?url";
 
-const HeaderLogo = () => {
+const MobileHeaderLogo = () => {
   return (
     <header className={styles.header}>
       <Image src={sampleImageSrc} alt="펫그룹 이미지" width={40} height={40} />
@@ -10,4 +10,4 @@ const HeaderLogo = () => {
   );
 };
 
-export default HeaderLogo;
+export default MobileHeaderLogo;
